misc: Make it stable compatible
This commit adds a new feature "nightly" to hides extern_types
requirement behind a cfg!.

When the feature isn't active, a repr(C) struct is defined with a size
of 0 (with one of the old tricks around for that matter)

I took the liberty to hide this boilerplate around a simple macro.

This fix build on stable compiler.